% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnehotEncoding.R
\name{oneHotEncoding}
\alias{oneHotEncoding}
\title{One hot encoding
One hot encoding function is used to convert the factor variables into multiple columns
based on the unique values}
\usage{
oneHotEncoding(data)
}
\arguments{
\item{data}{frame is taken}
}
\value{
data frame with the one hot encoding applied to all the variables with type as factor
}
\description{
One hot encoding
One hot encoding function is used to convert the factor variables into multiple columns
based on the unique values
}

